app/agents/quantum_nlp_agent.py
from app.agents.base import Agent
from app.chat_with_ollama import ChatGPT
from app.agents.skill_manager import SkillManager
from typing import Dict, Any, List
import logging
import json
from tenacity import retry, stop_after_attempt, wait_exponential

logger = logging.getLogger(__name__)

class QuantumNLPAgent(Agent):
    def __init__(self, agent_id: str, name: str, skill_manager: SkillManager, llm: ChatGPT):
        super().__init__(agent_id, name, skill_manager, llm)

    async def process_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        try:
            parsed_task = await self._parse_task(task['content'])
            embeddings = await self._generate_quantum_embeddings(parsed_task)
            optimized_tasks = await self._optimize_task_order(embeddings)
            result = await self._process_optimized_tasks(optimized_tasks)
            return {"result": result}
        except Exception as e:
            logger.error(f"Error processing task: {str(e)}")
            return {"error": str(e)}

    # Implement other methods (_parse_task, _generate_quantum_embeddings, _optimize_task_order, _process_optimized_tasks)